Move Managers to pootle_store.managers
also moves SuggestionStates into pootle_store.util to avoid cyclic imports
